added planet support and auctionhouse support

+++ src/main/java/com/yolosec/data/ItemDAO.java
@@ -1,9 +1,3 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
 package com.yolosec.data;
 
 import com.ptsesd.groepb.shared.ItemComm;
@@ -19,9 +13,9 @@ public interface ItemDAO {
     
     public List<ItemComm> getAuctionHouse();
     
-    public void buyItem(int spaceshipId, int itemId);
+    public void buyItem(int spaceshipId, long itemId);
     
-    public void sellItem(int spaceshipId, int itemId);
+    public void sellItem(int spaceshipId, long itemId);
     
-    public void cancelAuction(int spaceshipId, int itemId);
+    public void cancelAuction(int spaceshipId, long itemId);
 }
